+++ packages/synapse-interface/pages/index.tsx
@@ -4,8 +4,8 @@ import { LandingPageWrapper } from '@/components/layouts/LandingPageWrapper'
 import ReactGA from 'react-ga'
 import useSyncQueryParamsWithBridgeState from '@/utils/hooks/useSyncQueryParamsWithBridgeState'
 import { EthDencunUpgradeBanner } from '@/components/Maintenance/EthDencunUpgrade'
-import { EcotoneForkUpgradeBanner } from '@/components/Maintenance/EcotoneForkUpgrade'
-import { MetisDowntimeBanner } from '@/components/Maintenance/MetisUpgrade'
+import { EcotoneForkUpgradeBanner } from '@/components/Maintenance/Events/EcotoneForkUpgrade'
+import { MetisDowntimeBanner } from '@/components/Maintenance/Events/MetisUpgrade'
 
 // TODO: someone should add this to the .env, disable if blank, etc.
 // this is being added as a hotfix to assess user load on the synapse explorer api
